Fix TensorBoard component
