+++ contrib/submit-job-v2/src/App/ProtocolForm.tsx
@@ -129,24 +129,7 @@ const cx = classNames.bind(styles);
 
 initializeIcons();
 
-interface ILogDirectoryObj {
-  [key: string]: string;
-}
-
-interface IStorage {
-  type: string;
-  hostIP: string;
-  port: string;
-  remotePath: string;
-}
-
-interface ITensorBoardConfig {
-  randomStr: string;
-  storage: IStorage;
-  logDirectories: ILogDirectoryObj;
-}
-
-interface IParameterObj {
+interface IArrayObj {
   [key: string]: string;
 }
 
@@ -177,8 +160,6 @@ interface IProtocolState {
   loading: boolean;
   showParameters: boolean;
   showEditor: boolean;
-  tensorBoardConfig: ITensorBoardConfig;
-  enableTensorBoard: boolean;
 }
 
 export default class ProtocolForm extends React.Component<IProtocolProps, IProtocolState> {
@@ -354,8 +335,8 @@ export default class ProtocolForm extends React.Component<IProtocolProps, IProto
             </Stack>
             <Stack className={styles.item}>
               <TensorBoard
-                setTensorBoardConfig={this.setTensorBoardConfig}
-                toggleTensorBoard={this.toggleTensorBoard}
+                protocol={this.state.protocol}
+                setProtocol={this.setProtocol}
               />
             </Stack>
             <Stack className={styles.item}>
@@ -399,39 +380,17 @@ export default class ProtocolForm extends React.Component<IProtocolProps, IProto
         }
         protocol.name = this.getCloneJobName(source.jobName);
       }
-      let enableTensorBoard = false;
-      let tensorBoardConfig = {
-        randomStr: "",
-        storage: {
-          type: "HDFS",
-          hostIP: "",
-          port: "",
-          remotePath: "",
-        },
-        logDirectories: {
-          default: "$TB_ROOT",
-        },
-      };
-      if (protocol !== null) {
-        if (Object.prototype.hasOwnProperty.call(protocol, "extras")) {
-          if (Object.prototype.hasOwnProperty.call(protocol.extras, "tensorBoard")) {
-            enableTensorBoard = true;
-            tensorBoardConfig = protocol.extras.tensorBoard;
-          }
-        }
-      }
       if (protocol) {
         this.setState({
           jobName: protocol.name,
           protocol,
           protocolYAML: yaml.safeDump(protocol),
-          enableTensorBoard,
-          tensorBoardConfig,
         });
       }
     } catch (err) {
       alert(err.message);
     }
+    console.log(this.state.enableTensorBoard);
     this.setState({ loading: false });
   }
 
@@ -454,20 +413,13 @@ export default class ProtocolForm extends React.Component<IProtocolProps, IProto
     }
   }
 
-
-
   private onSelectProtocol = (text: string) => {
     try {
       const protocol = yaml.safeLoad(text);
-      let enableTensorBoard = false;
-      if (protocol.hasOwnProperty("extras") && protocol.extras.hasOwnProperty("tensorBoard")) {
-        enableTensorBoard = true;
-      }
       this.setState({
         jobName: protocol.name || "",
         protocol,
         protocolYAML: text,
-        enableTensorBoard,
       });
     } catch (err) {
       alert(err.message);
@@ -485,15 +437,10 @@ export default class ProtocolForm extends React.Component<IProtocolProps, IProto
       const text = fileReader.result as string;
       try {
         const protocol = yaml.safeLoad(text);
-        let enableTensorBoard = false;
-        if (protocol.hasOwnProperty("extras") && protocol.extras.hasOwnProperty("tensorBoard")) {
-          enableTensorBoard = true;
-        }
         this.setState({
           jobName: protocol.name || "",
           protocol,
           protocolYAML: text,
-          enableTensorBoard,
         });
       } catch (err) {
         alert(err.message);
@@ -524,7 +471,7 @@ export default class ProtocolForm extends React.Component<IProtocolProps, IProto
       const setParameter = (event: React.FormEvent<HTMLInputElement | HTMLTextAreaElement>, value?: string) => {
         if (value !== undefined) {
           const protocol = this.state.protocol;
-          (protocol.parameters as IParameterObj)[item.key] = value;
+          (protocol.parameters as IArrayObj)[item.key] = value;
           this.setState({
             protocol,
             protocolYAML: yaml.safeDump(protocol),
@@ -576,12 +523,6 @@ export default class ProtocolForm extends React.Component<IProtocolProps, IProto
 
   private openEditor = async (event: React.MouseEvent<HTMLButtonElement, MouseEvent>) => {
     event.preventDefault();
-    if (this.state.enableTensorBoard) {
-      await this.deleteTensorBoardConfig();
-      await this.addTensorBoardConfig();
-    } else {
-      await this.deleteTensorBoardConfig();
-    }
     this.setState({ showEditor: true });
   }
 
@@ -594,29 +535,10 @@ export default class ProtocolForm extends React.Component<IProtocolProps, IProto
     const text = this.state.protocolYAML;
     try {
       const protocol = yaml.safeLoad(text);
-      let enableTensorBoard = false;
-      let tensorBoardConfig = {
-        randomStr: "",
-        storage: {
-          type: "HDFS",
-          hostIP: "",
-          port: "",
-          remotePath: "",
-        },
-        logDirectories: {
-          default: "$TB_ROOT",
-        },
-      };
-      if (protocol.hasOwnProperty("extras") && protocol.extras.hasOwnProperty("tensorBoard")) {
-        enableTensorBoard = true;
-        tensorBoardConfig = protocol.extras.tensorBoard;
-      }
       this.setState({
         jobName: protocol.name || "",
         protocol,
         showEditor: false,
-        tensorBoardConfig,
-        enableTensorBoard,
       });
     } catch (err) {
       alert(err.message);
@@ -632,193 +554,7 @@ export default class ProtocolForm extends React.Component<IProtocolProps, IProto
     });
   }
 
-  private setTensorBoardConfig = (tensorBoardConfig: ITensorBoardConfig) => {
-    this.setState({ tensorBoardConfig });
-  }
-
-  public toggleTensorBoard = (checked: boolean) => {
-    this.setState({ enableTensorBoard: checked });
-  }
-
-  private generatePreCommand = () => {
-    let preCommand = [];
-    let mountPath = "/TMP_TENSORBOARD_LOG";
-    const storageConfig = this.state.tensorBoardConfig.storage;
-    preCommand.push("`#TENSORBOARD_STORAGE_START`");
-    preCommand.push("`#Auto generated code, please do not modify`");
-    preCommand.push(
-      `if [ ! -d ${mountPath} ]; then mkdir --parents ${
-      mountPath
-      }; fi`,
-    );
-    preCommand.push("apt-get update");
-    switch (storageConfig.type) {
-      case "HDFS":
-        preCommand.push([
-          "apt-get install -y git fuse golang",
-          "git clone --recursive https://github.com/Microsoft/hdfs-mount.git",
-          "cd hdfs-mount",
-          "make",
-          "cp hdfs-mount /bin",
-          "cd ..",
-          "rm -rf hdfs-mount",
-        ].join("&&"));
-        if (storageConfig.hostIP === undefined || storageConfig.remotePath === undefined) {
-          preCommand = [];
-          break;
-        }
-        preCommand.push(`(hdfs-mount ${storageConfig.hostIP}:${storageConfig.port} ${mountPath} &)`);
-        preCommand.push(`sleep 5`);
-        mountPath += storageConfig.remotePath;
-        break;
-      case "NFS":
-        if (storageConfig.hostIP === undefined || storageConfig.remotePath === undefined) {
-          preCommand = [];
-          break;
-        }
-        preCommand.push("apt-get install -y nfs-common");
-        preCommand.push(`mount -t nfs4 ${storageConfig.hostIP}:${storageConfig.remotePath} ${mountPath}`);
-        break;
-      default:
-        preCommand = [];
-        break;
-    }
-    if (preCommand.length === 0) {
-      alert("Please complete the storage config!");
-      return [];
-    }
-    preCommand.push(`export TB_ROOT=${mountPath}`);
-    preCommand.push("`#TENSORBOARD_STORAGE_END`");
-    return preCommand;
-  }
-
-  private injectCommand = () => {
-    const preCommand = this.generatePreCommand();
-    if (preCommand.length === 0) {
-      return false;
-    }
-    const protocol = yaml.safeLoad(this.state.protocolYAML);
-    if (protocol.hasOwnProperty("taskRoles")) {
-      const obj = protocol.taskRoles;
-      Object.keys(obj).forEach((key) => {
-        obj[key].commands = preCommand.concat(obj[key].commands);
-      });
-    }
-    this.setState({
-      protocol,
-      protocolYAML: yaml.safeDump(protocol),
-    });
-  }
-
-  private addTensorBoardConfig = () => {
-    const protocol = this.state.protocol;
-    if (protocol === undefined) {
-      return;
-    }
-    const tensorBoardConfig = this.state.tensorBoardConfig;
-    tensorBoardConfig.randomStr = Math.random().toString(36).slice(2, 10);
-    if (Object.prototype.hasOwnProperty.call(protocol, "extras")) {
-      protocol.extras.tensorBoard = tensorBoardConfig;
-    } else {
-      protocol.extras = { tensorBoard: tensorBoardConfig };
-    }
-    this.setState({
-      protocol,
-      protocolYAML: yaml.safeDump(protocol),
-      tensorBoardConfig,
-    });
-    this.addTensorBoardTaskRole();
-    this.injectCommand();
-  }
-
-  private addTensorBoardTaskRole = () => {
-    const protocol = this.state.protocol;
-    const tensorBoardConfig = this.state.tensorBoardConfig;
-    const tensorBoardName = `TensorBoard_${tensorBoardConfig.randomStr}`;
-    const tensorBoardImage = `tensorBoardImage_${tensorBoardConfig.randomStr}`;
-    const tensorBoardPort = `tensorBoardPort_${tensorBoardConfig.randomStr}`;
-
-    if (!Object.prototype.hasOwnProperty.call(protocol, "prerequisites")) {
-      protocol.prerequisites = [];
-    }
-    protocol.prerequisites = update(protocol.prerequisites, {
-      $push: [{
-        protocolVersion: 2,
-        name: tensorBoardImage,
-        type: "dockerimage",
-        version: "1.0 - r1.4",
-        contributor: "OpenPAI",
-        uri: "openpai/pai.example.tensorflow",
-      }],
-    });
-
-    const portList = ` --port=$PAI_CONTAINER_HOST_${tensorBoardPort}_PORT_LIST`;
-    const logPathList: string[] = [];
-    const logDirectories = tensorBoardConfig.logDirectories as IParameterObj;
-    Object.keys(logDirectories).forEach((key) => {
-      logPathList.push(`${key}:${logDirectories[key]}`);
-    });
-    const logPath = logPathList.join(":");
-
-    if (!Object.prototype.hasOwnProperty.call(protocol, "taskRoles")) {
-      protocol.taskRoles = {};
-    }
-    const minSucceededInstances = Object.getOwnPropertyNames(protocol.taskRoles).length;
-    protocol.taskRoles[tensorBoardName] = {
-      instances: 1,
-      completion: {
-        minFailedInstances: 1,
-        minSucceededInstances: minSucceededInstances <= 0 ? null : minSucceededInstances,
-      },
-      taskRetryCount: 0,
-      dockerImage: tensorBoardImage,
-      resourcePerInstance:
-      {
-        cpu: 4,
-        memoryMB: 8192,
-        gpu: 0,
-        ports: {},
-      },
-      commands: [`tensorboard --logdir=${logPath} ${portList}`],
-    };
-    protocol.taskRoles[tensorBoardName].resourcePerInstance.ports[tensorBoardPort] = 1;
-
-    this.setState({
-      protocol,
-      protocolYAML: yaml.safeDump(protocol),
-    });
-  }
-
-  private deleteTensorBoardConfig = () => {
-    const protocol = this.state.protocol;
-    const randomStr = this.state.tensorBoardConfig.randomStr;
-    const tensorBoardName = `TensorBoard_${randomStr}`;
-    const tensorBoardImage = `tensorBoardImage_${randomStr}`;
-    if (Object.prototype.hasOwnProperty.call(protocol, "taskRoles")) {
-      delete protocol.taskRoles[tensorBoardName];
-      Object.keys(protocol.taskRoles).forEach((key) => {
-        const startIndex = protocol.taskRoles[key].commands.indexOf("`#TENSORBOARD_STORAGE_START`");
-        const endIndex = protocol.taskRoles[key].commands.indexOf("`#TENSORBOARD_STORAGE_END`");
-        if (startIndex !== -1 && endIndex !== -1) {
-          protocol.taskRoles[key].commands.splice(startIndex, endIndex - startIndex + 1);
-        }
-      });
-    }
-    if (Object.prototype.hasOwnProperty.call(protocol, "prerequisites")) {
-      const index = protocol.prerequisites.findIndex((ele: any) => ele.name === tensorBoardImage);
-      if (index !== -1) {
-        protocol.prerequisites.splice(index, 1);
-      }
-    }
-    if (Object.prototype.hasOwnProperty.call(protocol, "extras")) {
-      if (Object.prototype.hasOwnProperty.call(protocol.extras, "tensorBoard")) {
-        delete protocol.extras.tensorBoard;
-      }
-      if (Object.getOwnPropertyNames(protocol.extras).length === 0) {
-        delete protocol.extras;
-      }
-    }
-
+  private setProtocol = (protocol: any) => {
     this.setState({
       protocol,
       protocolYAML: yaml.safeDump(protocol),
@@ -830,9 +566,15 @@ export default class ProtocolForm extends React.Component<IProtocolProps, IProto
     if (!this.state.protocolYAML) {
       return;
     }
-    await this.deleteTensorBoardConfig();
-    if (this.state.enableTensorBoard) {
-      const storage = this.state.tensorBoardConfig.storage;
+    const protocol = yaml.safeLoad(this.state.protocolYAML);
+    let tensorBoardConfig = undefined;
+    if (Object.prototype.hasOwnProperty.call(protocol, "extras")) {
+      if (Object.prototype.hasOwnProperty.call(protocol.extras, "tensorBoard")) {
+        tensorBoardConfig = protocol.extras.tensorBoard;
+      }
+    }
+    if (tensorBoardConfig !== undefined) {
+      const storage = tensorBoardConfig.storage;
       switch (storage.type) {
         case "HDFS":
           if (storage.hostIP === "" || storage.port === "" || storage.remotePath === "") {
@@ -850,9 +592,7 @@ export default class ProtocolForm extends React.Component<IProtocolProps, IProto
           alert("Unsupport external storage type!");
           return;
       }
-      await this.addTensorBoardConfig();
     }
-    const protocol = yaml.safeLoad(this.state.protocolYAML);
     if (protocol.hasOwnProperty("extras")) {
       protocol.extras.submitFrom = this.props.pluginId;
     } else {
