+++ contrib/submit-job-v2/src/App/TensorBoard.tsx
@@ -1,10 +1,11 @@
 import React from "react";
 import {
   ComboBox, IComboBox, IComboBoxOption, Toggle, Callout, DefaultButton, Dropdown, DropdownMenuItemType,
-  Fabric, IDropdownOption, Stack, PrimaryButton, TextField, mergeStyleSets, enableBodyScroll,
+  Fabric, IDropdownOption, Stack, PrimaryButton, TextField, mergeStyleSets, enableBodyScroll, TooltipHostBase,
 } from "office-ui-fabric-react";
+import update from "immutability-helper";
 
-interface ILogDirectoryObj {
+interface IArrayObj {
   [key: string]: string;
 }
 
@@ -18,17 +19,16 @@ interface IStorage {
 interface ITensorBoardConfig {
   randomStr: string;
   storage: IStorage;
-  logDirectories: ILogDirectoryObj;
+  logDirectories: IArrayObj;
 }
 
 interface ITensorBoardProps {
-  externalStorageTypeOptions: IComboBoxOption[];
-  enableTensorBoard: boolean;
-  setTensorBoardConfig: ((tensorBoardConfig: ITensorBoardConfig) => void);
-  toggleTensorBoard: ((checked: boolean) => void);
+  protocol: any;
+  setProtocol: ((protocol: any) => void);
 }
 
 interface ITensorBoardState {
+  protocol: any;
   tensorBoardConfig: ITensorBoardConfig,
   enableTensorBoard: boolean,
 }
@@ -40,11 +40,11 @@ const defaultExternalStorageTypeOptions: IComboBoxOption[] = [
 
 export default class TensorBoard extends React.Component<ITensorBoardProps, ITensorBoardState> {
   public static defaultProps: Partial<ITensorBoardProps> = {
-    enableTensorBoard: false,
-    externalStorageTypeOptions: defaultExternalStorageTypeOptions,
+    protocol: Object.create(null),
   };
 
   public state = {
+    protocol: this.props.protocol,
     tensorBoardConfig: {
       randomStr: "",
       storage: {
@@ -60,6 +60,77 @@ export default class TensorBoard extends React.Component<ITensorBoardProps, ITen
     enableTensorBoard: false,
     externalStorageTypeOptions: defaultExternalStorageTypeOptions,
   };
+
+  public componentWillMount() {
+    const protocol = this.props.protocol;
+    let tensorBoardConfig = {
+      randomStr: "",
+      storage: {
+        type: "HDFS",
+        hostIP: "",
+        port: "",
+        remotePath: "",
+      },
+      logDirectories: {
+        default: "$TB_ROOT",
+      },
+    };
+    let enableTensorBoard = false;
+    if (protocol !== null) {
+      if (Object.prototype.hasOwnProperty.call(protocol, "extras")) {
+        if (Object.prototype.hasOwnProperty.call(protocol.extras, "tensorBoard")) {
+          tensorBoardConfig = protocol.extras.tensorBoard;
+          enableTensorBoard = true;
+        }
+      }
+    }
+    this.setState(({
+      tensorBoardConfig,
+      protocol,
+      enableTensorBoard,
+    }),
+      () => {
+        this.updateProtocol();
+      },
+    );
+  }
+
+  public componentWillReceiveProps(nextProps: ITensorBoardProps) {
+    const { protocol } = nextProps;
+    let tensorBoardConfig = {
+      randomStr: "",
+      storage: {
+        type: "HDFS",
+        hostIP: "",
+        port: "",
+        remotePath: "",
+      },
+      logDirectories: {
+        default: "$TB_ROOT",
+      },
+    };
+    let enableTensorBoard = false;
+    if (protocol !== null) {
+      if (Object.prototype.hasOwnProperty.call(protocol, "extras")) {
+        if (Object.prototype.hasOwnProperty.call(protocol.extras, "tensorBoard")) {
+          tensorBoardConfig = protocol.extras.tensorBoard;
+          enableTensorBoard = true;
+        }
+      }
+    }
+    if (protocol != this.state.protocol) {
+      this.setState(({
+        tensorBoardConfig,
+        protocol,
+        enableTensorBoard,
+      }),
+        () => {
+          this.updateProtocol();
+        },
+      );
+    }
+
+  }
 
   public render() {
     return (
@@ -105,7 +176,6 @@ export default class TensorBoard extends React.Component<ITensorBoardProps, ITen
             />
             <TextField
               label="Port"
-              placeholder={`9000`}
               value={this.state.tensorBoardConfig.storage.port}
               onChange={this.setPort}
               required={true}
@@ -143,7 +213,7 @@ export default class TensorBoard extends React.Component<ITensorBoardProps, ITen
   private renderTensorBoardLogPath = () => {
     if (this.state.enableTensorBoard) {
       const logPathList: string[] = [];
-      const logDirectories = this.state.tensorBoardConfig.logDirectories as ILogDirectoryObj;
+      const logDirectories = this.state.tensorBoardConfig.logDirectories as IArrayObj;
       Object.keys(logDirectories).forEach((key) => {
         logPathList.push(`${key}:${logDirectories[key]}`);
       });
@@ -160,8 +230,10 @@ export default class TensorBoard extends React.Component<ITensorBoardProps, ITen
 
   private toggleTensorBoard = (event: React.MouseEvent<HTMLElement, MouseEvent>, checked?: boolean) => {
     if (checked !== undefined) {
-      this.setState({ enableTensorBoard: checked });
-      this.props.toggleTensorBoard(checked);
+      this.setState(
+        ({ enableTensorBoard: checked }),
+        () => { this.updateProtocol(); },
+      );
     }
   }
 
@@ -169,8 +241,10 @@ export default class TensorBoard extends React.Component<ITensorBoardProps, ITen
     if (hostIP !== undefined) {
       const tensorBoardConfig = this.state.tensorBoardConfig;
       tensorBoardConfig.storage.hostIP = hostIP;
-      this.setState({ tensorBoardConfig });
-      this.props.setTensorBoardConfig(tensorBoardConfig);
+      this.setState(
+        ({ tensorBoardConfig }),
+        () => { this.updateProtocol(); },
+      );
     }
   }
 
@@ -178,8 +252,10 @@ export default class TensorBoard extends React.Component<ITensorBoardProps, ITen
     if (port !== undefined) {
       const tensorBoardConfig = this.state.tensorBoardConfig;
       tensorBoardConfig.storage.port = port;
-      this.setState({ tensorBoardConfig });
-      this.props.setTensorBoardConfig(tensorBoardConfig);
+      this.setState(
+        ({ tensorBoardConfig }),
+        () => { this.updateProtocol(); },
+      );
     }
   }
 
@@ -187,8 +263,10 @@ export default class TensorBoard extends React.Component<ITensorBoardProps, ITen
     if (remotePath !== undefined) {
       const tensorBoardConfig = this.state.tensorBoardConfig;
       tensorBoardConfig.storage.remotePath = remotePath;
-      this.setState({ tensorBoardConfig });
-      this.props.setTensorBoardConfig(tensorBoardConfig);
+      this.setState(
+        ({ tensorBoardConfig }),
+        () => { this.updateProtocol(); },
+      );
     }
   }
 
@@ -196,8 +274,198 @@ export default class TensorBoard extends React.Component<ITensorBoardProps, ITen
     if (option !== undefined) {
       const tensorBoardConfig = this.state.tensorBoardConfig;
       tensorBoardConfig.storage.type = option.key.toString();
-      this.setState({ tensorBoardConfig });
-      this.props.setTensorBoardConfig(tensorBoardConfig);
+      this.setState(
+        ({ tensorBoardConfig }),
+        () => { this.updateProtocol(); },
+      );
     }
+  }
+
+  private generatePreCommand = () => {
+    let preCommand = [];
+    let mountPath = "/TMP_TENSORBOARD_LOG";
+    const storageConfig = this.state.tensorBoardConfig.storage;
+    preCommand.push("`#TENSORBOARD_STORAGE_START`");
+    preCommand.push("`#Auto generated code, please do not modify`");
+    preCommand.push(
+      `if [ ! -d ${mountPath} ]; then mkdir --parents ${
+      mountPath
+      }; fi`,
+    );
+    preCommand.push("apt-get update");
+    switch (storageConfig.type) {
+      case "HDFS":
+        preCommand.push([
+          "apt-get install -y git fuse golang",
+          "git clone --recursive https://github.com/Microsoft/hdfs-mount.git",
+          "cd hdfs-mount",
+          "make",
+          "cp hdfs-mount /bin",
+          "cd ..",
+          "rm -rf hdfs-mount",
+        ].join("&&"));
+        if (storageConfig.hostIP === undefined || storageConfig.remotePath === undefined) {
+          preCommand = [];
+          break;
+        }
+        preCommand.push(`(hdfs-mount ${storageConfig.hostIP}:${storageConfig.port} ${mountPath} &)`);
+        preCommand.push(`sleep 5`);
+        mountPath += storageConfig.remotePath;
+        break;
+      case "NFS":
+        if (storageConfig.hostIP === undefined || storageConfig.remotePath === undefined) {
+          preCommand = [];
+          break;
+        }
+        preCommand.push("apt-get install -y nfs-common");
+        preCommand.push(`mount -t nfs4 ${storageConfig.hostIP}:${storageConfig.remotePath} ${mountPath}`);
+        break;
+      default:
+        preCommand = [];
+        break;
+    }
+    if (preCommand.length === 0) {
+      alert("Please complete the storage config!");
+      return [];
+    }
+    preCommand.push(`export TB_ROOT=${mountPath}`);
+    preCommand.push("`#TENSORBOARD_STORAGE_END`");
+    return preCommand;
+  }
+
+  private injectCommand = () => {
+    const preCommand = this.generatePreCommand();
+    if (preCommand.length === 0) {
+      return false;
+    }
+    const protocol = this.state.protocol;
+    if (Object.prototype.hasOwnProperty.call(protocol, "taskRoles")) {
+      const obj = protocol.taskRoles;
+      Object.keys(obj).forEach((key) => {
+        obj[key].commands = preCommand.concat(obj[key].commands);
+      });
+      this.setState({ protocol });
+    }
+
+  }
+
+  private addTensorBoardConfig = () => {
+    const protocol = this.state.protocol;
+    if (protocol === undefined) {
+      return;
+    }
+    const tensorBoardConfig = this.state.tensorBoardConfig;
+    tensorBoardConfig.randomStr = Math.random().toString(36).slice(2, 10);
+    if (Object.prototype.hasOwnProperty.call(protocol, "extras")) {
+      protocol.extras.tensorBoard = tensorBoardConfig;
+    } else {
+      protocol.extras = { tensorBoard: tensorBoardConfig };
+    }
+    this.setState({
+      protocol,
+      tensorBoardConfig,
+    });
+    this.addTensorBoardTaskRole();
+    this.injectCommand();
+  }
+
+  private addTensorBoardTaskRole = () => {
+    const protocol = this.state.protocol;
+    const tensorBoardConfig = this.state.tensorBoardConfig;
+    const tensorBoardName = `TensorBoard_${tensorBoardConfig.randomStr}`;
+    const tensorBoardImage = `tensorBoardImage_${tensorBoardConfig.randomStr}`;
+    const tensorBoardPort = `tensorBoardPort_${tensorBoardConfig.randomStr}`;
+
+    if (!Object.prototype.hasOwnProperty.call(protocol, "prerequisites")) {
+      protocol.prerequisites = [];
+    }
+    protocol.prerequisites = update(protocol.prerequisites, {
+      $push: [{
+        protocolVersion: 2,
+        name: tensorBoardImage,
+        type: "dockerimage",
+        version: "1.0 - r1.4",
+        contributor: "OpenPAI",
+        uri: "openpai/pai.example.tensorflow",
+      }],
+    });
+
+    const portList = ` --port=$PAI_CONTAINER_HOST_${tensorBoardPort}_PORT_LIST`;
+    const logPathList: string[] = [];
+    const logDirectories = tensorBoardConfig.logDirectories as IArrayObj;
+    Object.keys(logDirectories).forEach((key) => {
+      logPathList.push(`${key}:${logDirectories[key]}`);
+    });
+    const logPath = logPathList.join(":");
+
+    if (!Object.prototype.hasOwnProperty.call(protocol, "taskRoles")) {
+      protocol.taskRoles = {};
+    }
+    const minSucceededInstances = Object.getOwnPropertyNames(protocol.taskRoles).length;
+    protocol.taskRoles[tensorBoardName] = {
+      instances: 1,
+      completion: {
+        minFailedInstances: 1,
+        minSucceededInstances: minSucceededInstances <= 0 ? null : minSucceededInstances,
+      },
+      taskRetryCount: 0,
+      dockerImage: tensorBoardImage,
+      resourcePerInstance:
+      {
+        cpu: 4,
+        memoryMB: 8192,
+        gpu: 0,
+        ports: {},
+      },
+      commands: [`tensorboard --logdir=${logPath} ${portList}`],
+    };
+    protocol.taskRoles[tensorBoardName].resourcePerInstance.ports[tensorBoardPort] = 1;
+    this.setState({ protocol });
+  }
+
+  private deleteTensorBoardConfig = () => {
+    const protocol = this.state.protocol;
+    const randomStr = this.state.tensorBoardConfig.randomStr;
+    const tensorBoardName = `TensorBoard_${randomStr}`;
+    const tensorBoardImage = `tensorBoardImage_${randomStr}`;
+    if (Object.prototype.hasOwnProperty.call(protocol, "taskRoles")) {
+      delete protocol.taskRoles[tensorBoardName];
+      Object.keys(protocol.taskRoles).forEach((key) => {
+        const startIndex = protocol.taskRoles[key].commands.indexOf("`#TENSORBOARD_STORAGE_START`");
+        const endIndex = protocol.taskRoles[key].commands.indexOf("`#TENSORBOARD_STORAGE_END`");
+        if (startIndex !== -1 && endIndex !== -1) {
+          protocol.taskRoles[key].commands.splice(startIndex, endIndex - startIndex + 1);
+        }
+      });
+      if (Object.getOwnPropertyNames(protocol.taskRoles).length === 0) {
+        delete protocol.taskRoles;
+      }
+    }
+    if (Object.prototype.hasOwnProperty.call(protocol, "prerequisites")) {
+      const index = protocol.prerequisites.findIndex((ele: any) => ele.name === tensorBoardImage);
+      if (index !== -1) {
+        protocol.prerequisites.splice(index, 1);
+      }
+      if (protocol.prerequisites.length === 0) {
+        delete protocol.prerequisites;
+      }
+    }
+    if (Object.prototype.hasOwnProperty.call(protocol, "extras")) {
+      if (Object.prototype.hasOwnProperty.call(protocol.extras, "tensorBoard")) {
+        delete protocol.extras.tensorBoard;
+      }
+      if (Object.getOwnPropertyNames(protocol.extras).length === 0) {
+        delete protocol.extras;
+      }
+    }
+    this.setState({ protocol });
+  }
+
+  private updateProtocol = () => {
+    this.deleteTensorBoardConfig();
+    if (this.state.enableTensorBoard) {
+      this.addTensorBoardConfig();
+    }
+    this.props.setProtocol(this.state.protocol);
   }
 }
